fix: candy machine slugs (#192)
* Fix slugs

* Change urls to new slugs
